Add simple version of orca_read module
Reading orca output is set by "-in orca_cis" and giving the path to a
molden file generated by orca_2mkl with a name of form base.molden*, the
CIS coefficients will then be read from the base.cis file.

Tested for ORCA 4.1.2 only for restricted caluclations and for singlet
excitations.